+++ tests/frontend/test_tdmprogram.py
@@ -835,3 +835,82 @@ class TestTDMValidation:
             CircuitError, match="Parameter has value '-999' while its valid range is "
         ):
             self.compile_test_program(device, args=args)
+
+
+class TestUnrolling:
+    """Test that (un)rolling programs work as expected."""
+
+    def test_unroll(self):
+        """Test unrolling program."""
+        n = 2
+
+        prog = tdmprogram.TDMProgram(N=2)
+        with prog.context([0] * n, [0] * n, [0] * n) as (p, q):
+            ops.Sgate(0.5643, 0) | q[1]
+            ops.BSgate(p[0]) | (q[1], q[0])
+            ops.Rgate(p[1]) | q[1]
+            ops.MeasureHomodyne(p[2]) | q[0]
+
+        prog_length = len(prog.circuit)
+        assert prog_length == 4
+
+        prog.unroll()
+        assert len(prog.circuit) == n * prog_length
+
+        prog.roll()
+        assert len(prog.circuit) == prog_length
+
+    def test_unroll_shots(self):
+        """Test unrolling program several times using different number of shots."""
+        n = 2
+        shots = 2
+
+        prog = tdmprogram.TDMProgram(N=2)
+        with prog.context([0] * n, [0] * n, [0] * n) as (p, q):
+            ops.Sgate(0.5643, 0) | q[1]
+            ops.BSgate(p[0]) | (q[1], q[0])
+            ops.Rgate(p[1]) | q[1]
+            ops.MeasureHomodyne(p[2]) | q[0]
+
+        prog_length = len(prog.circuit)
+        assert prog_length == 4
+
+        prog.unroll(shots=shots)
+        assert len(prog.circuit) == n * shots * prog_length
+
+        # unroll once more with the same number of shots to cover caching
+        prog.unroll(shots=shots)
+        assert len(prog.circuit) == n * shots * prog_length
+
+        # unroll once more with a different number of shots
+        shots = 3
+        prog.unroll(shots=shots)
+        assert len(prog.circuit) == n * shots * prog_length
+
+        prog.roll()
+        assert len(prog.circuit) == prog_length
+
+    @pytest.mark.parametrize(
+        "space_unrolled, start_locked", [[True, True], [True, False], [False, True], [False, False]]
+    )
+    def test_locking_when_unrolling(self, space_unrolled, start_locked):
+        """Test that a locked program can be (un)rolled with the locking intact."""
+        prog = tdmprogram.TDMProgram(N=2)
+
+        with prog.context([0, 0], [0, 0], [0, 0]) as (p, q):
+            ops.Sgate(0.5643, 0) | q[1]
+            ops.BSgate(p[0]) | (q[1], q[0])
+            ops.Rgate(p[1]) | q[1]
+            ops.MeasureHomodyne(p[2]) | q[0]
+
+        if start_locked:
+            prog.lock()
+
+        assert prog.locked == start_locked
+
+        if space_unrolled:
+            prog.space_unroll()
+        else:
+            prog.unroll()
+
+        assert prog.locked == start_locked