+++ .github/CHANGELOG.md
@@ -6,7 +6,19 @@
 
 ### Improvements
 
+* A locked program can now be (un)rolled, and automatically restores the lock if there.
+[(#703)](https://github.com/XanaduAI/strawberryfields/pull/703)
+
+* Rolling and unrolling now only happens in place, and does no longer return the (un)rolled circuit.
+[(#702)](https://github.com/XanaduAI/strawberryfields/pull/702)
+
 ### Bug fixes
+
+* Trying to unroll an already unrolled program with a different number of shots works as expected.
+[(#702)](https://github.com/XanaduAI/strawberryfields/pull/702)
+
+* Fixed bug with vacuum modes missing.
+[(#702)](https://github.com/XanaduAI/strawberryfields/pull/702)
 
 ### Documentation
 
@@ -18,7 +30,7 @@
 
 This release contains contributions from (in alphabetical order):
 
-Mikhail Andrenkov
+Mikhail Andrenkov, Theodor Isacsson
 
 ---